remove Redis (Heroku) from project (simplify for beginners!) see: https://github.com/dwyl/hits/issues/54
