+++ lib/extract_request_data.js
@@ -13,8 +13,8 @@ module.exports = function extract (request) {
 
   // get url the client requested:
   var url = request.url.replace('.svg', '')
-          .replace('.png', '')
-          .replace('https://github.com/', ''); // strip to save storage
+    .replace('.png', '')
+    .replace('https://github.com/', ''); // strip to save storage
 
   if(h['accept-language']) { // Language for: github.com/dwyl/hits/issues/43
     if (h['accept-language'].indexOf(',') > -1) { // e.g: en-GB,en;q=0.5
@@ -25,5 +25,5 @@ module.exports = function extract (request) {
   }
 
   return [ Math.floor(Date.now()/1000), 
-      url, h['user-agent'], ip, lang ].join('|');
+    url, h['user-agent'], ip, lang ].join('|');
 }
