+++ lib/hits.js
@@ -4,10 +4,10 @@
  * @param {Function} callback - call this once redis responds
  */
 module.exports = function add (hit, callback) {
-  if(process.env.REDISCLOUD_URL) {
-    return require('./db_redis.js')(hit, callback);
-  }
-  else {
+  // if(process.env.REDISCLOUD_URL) {
+  //   return require('./db_redis.js')(hit, callback);
+  // }
+  // else {
     return require('./db_filesystem.js')(hit, callback);
-  }
+  // }
 }
